requirements.txt
flask==2.3.3
gunicorn==21.2.0
python-dotenv==1.0.0
Werkzeug==2.3.7
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7 